refactor: move lax_numpy window functions into their own file
