+++ jax/_src/numpy/lax_numpy.py
@@ -11544,164 +11544,6 @@ def _canonicalize_tuple_index(arr_ndim, idx):
     idx = tuple(idx) + colons
   return idx
 
-
-@export
-def blackman(M: int) -> Array:
-  """Return a Blackman window of size M.
-
-  JAX implementation of :func:`numpy.blackman`.
-
-  Args:
-    M: The window size.
-
-  Returns:
-    An array of size M containing the Blackman window.
-
-  Examples:
-    >>> with jnp.printoptions(precision=2, suppress=True):
-    ...   print(jnp.blackman(4))
-    [-0.    0.63  0.63 -0.  ]
-
-  See also:
-    - :func:`jax.numpy.bartlett`: return a Bartlett window of size M.
-    - :func:`jax.numpy.hamming`: return a Hamming window of size M.
-    - :func:`jax.numpy.hanning`: return a Hanning window of size M.
-    - :func:`jax.numpy.kaiser`: return a Kaiser window of size M.
-  """
-  M = core.concrete_or_error(int, M, "M argument of jnp.blackman")
-  dtype = dtypes.canonicalize_dtype(dtypes.float_)
-  if M <= 1:
-    return ones(M, dtype)
-  n = lax.iota(dtype, M)
-  return 0.42 - 0.5 * ufuncs.cos(2 * pi * n / (M - 1)) + 0.08 * ufuncs.cos(4 * pi * n / (M - 1))
-
-
-@export
-def bartlett(M: int) -> Array:
-  """Return a Bartlett window of size M.
-
-  JAX implementation of :func:`numpy.bartlett`.
-
-  Args:
-    M: The window size.
-
-  Returns:
-    An array of size M containing the Bartlett window.
-
-  Examples:
-    >>> with jnp.printoptions(precision=2, suppress=True):
-    ...   print(jnp.bartlett(4))
-    [0.   0.67 0.67 0.  ]
-
-  See also:
-    - :func:`jax.numpy.blackman`: return a Blackman window of size M.
-    - :func:`jax.numpy.hamming`: return a Hamming window of size M.
-    - :func:`jax.numpy.hanning`: return a Hanning window of size M.
-    - :func:`jax.numpy.kaiser`: return a Kaiser window of size M.
-  """
-  M = core.concrete_or_error(int, M, "M argument of jnp.bartlett")
-  dtype = dtypes.canonicalize_dtype(dtypes.float_)
-  if M <= 1:
-    return ones(M, dtype)
-  n = lax.iota(dtype, M)
-  return 1 - ufuncs.abs(2 * n + 1 - M) / (M - 1)
-
-
-@export
-def hamming(M: int) -> Array:
-  """Return a Hamming window of size M.
-
-  JAX implementation of :func:`numpy.hamming`.
-
-  Args:
-    M: The window size.
-
-  Returns:
-    An array of size M containing the Hamming window.
-
-  Examples:
-    >>> with jnp.printoptions(precision=2, suppress=True):
-    ...   print(jnp.hamming(4))
-    [0.08 0.77 0.77 0.08]
-
-  See also:
-    - :func:`jax.numpy.bartlett`: return a Bartlett window of size M.
-    - :func:`jax.numpy.blackman`: return a Blackman window of size M.
-    - :func:`jax.numpy.hanning`: return a Hanning window of size M.
-    - :func:`jax.numpy.kaiser`: return a Kaiser window of size M.
-  """
-  M = core.concrete_or_error(int, M, "M argument of jnp.hamming")
-  dtype = dtypes.canonicalize_dtype(dtypes.float_)
-  if M <= 1:
-    return ones(M, dtype)
-  n = lax.iota(dtype, M)
-  return 0.54 - 0.46 * ufuncs.cos(2 * pi * n / (M - 1))
-
-
-@export
-def hanning(M: int) -> Array:
-  """Return a Hanning window of size M.
-
-  JAX implementation of :func:`numpy.hanning`.
-
-  Args:
-    M: The window size.
-
-  Returns:
-    An array of size M containing the Hanning window.
-
-  Examples:
-    >>> with jnp.printoptions(precision=2, suppress=True):
-    ...   print(jnp.hanning(4))
-    [0.   0.75 0.75 0.  ]
-
-  See also:
-    - :func:`jax.numpy.bartlett`: return a Bartlett window of size M.
-    - :func:`jax.numpy.blackman`: return a Blackman window of size M.
-    - :func:`jax.numpy.hamming`: return a Hamming window of size M.
-    - :func:`jax.numpy.kaiser`: return a Kaiser window of size M.
-  """
-  M = core.concrete_or_error(int, M, "M argument of jnp.hanning")
-  dtype = dtypes.canonicalize_dtype(dtypes.float_)
-  if M <= 1:
-    return ones(M, dtype)
-  n = lax.iota(dtype, M)
-  return 0.5 * (1 - ufuncs.cos(2 * pi * n / (M - 1)))
-
-
-@export
-def kaiser(M: int, beta: ArrayLike) -> Array:
-  """Return a Kaiser window of size M.
-
-  JAX implementation of :func:`numpy.kaiser`.
-
-  Args:
-    M: The window size.
-    beta: The Kaiser window parameter.
-
-  Returns:
-    An array of size M containing the Kaiser window.
-
-  Examples:
-    >>> with jnp.printoptions(precision=2, suppress=True):
-    ...   print(jnp.kaiser(4, 1.5))
-    [0.61 0.95 0.95 0.61]
-
-  See also:
-    - :func:`jax.numpy.bartlett`: return a Bartlett window of size M.
-    - :func:`jax.numpy.blackman`: return a Blackman window of size M.
-    - :func:`jax.numpy.hamming`: return a Hamming window of size M.
-    - :func:`jax.numpy.hanning`: return a Hanning window of size M.
-  """
-  M = core.concrete_or_error(int, M, "M argument of jnp.kaiser")
-  dtype = dtypes.canonicalize_dtype(dtypes.float_)
-  if M <= 1:
-    return ones(M, dtype)
-  n = lax.iota(dtype, M)
-  alpha = 0.5 * (M - 1)
-  return i0(beta * ufuncs.sqrt(1 - ((n - alpha) / alpha) ** 2)) / i0(beta)
-
-
 def _gcd_cond_fn(xs: tuple[Array, Array]) -> Array:
   x1, x2 = xs
   return reductions.any(x2 != 0)
